requirements.txt
opencv-python
numpy
torch
torchvision
ultralytics
deep-sort-realtime
tqdm
Pillow
pyyaml
scipy
scikit-learn
matplotlib
pandas
